Factor artifact handling to ./artifacts
