+++ workers/generic-worker/artifacts_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/taskcluster/slugid-go/slugid"
 	tcclient "github.com/taskcluster/taskcluster/v44/clients/client-go"
 	"github.com/taskcluster/taskcluster/v44/clients/client-go/tcqueue"
+	"github.com/taskcluster/taskcluster/v44/workers/generic-worker/artifacts"
 )
 
 var (
@@ -22,7 +23,7 @@ var (
 func validateArtifacts(
 	t *testing.T,
 	payloadArtifacts []Artifact,
-	expected []TaskArtifact) {
+	expected []artifacts.TaskArtifact) {
 
 	// to test, create a dummy task run with given artifacts
 	// and then call Artifacts() method to see what
@@ -38,10 +39,18 @@ func validateArtifacts(
 	for i := range payloadArtifacts {
 		tr.Payload.Artifacts = append(tr.Payload.Artifacts, payloadArtifacts[i])
 	}
-	artifacts := tr.PayloadArtifacts()
+	got := tr.PayloadArtifacts()
 
-	if !reflect.DeepEqual(artifacts, expected) {
-		t.Fatalf("Expected different artifacts to be generated...\nExpected:\n%q\nActual:\n%q", expected, artifacts)
+	// RawContentFile contains a full path that depends on the specific
+	// task directory, so empty that value out before comparing.
+	for _, a := range got {
+		if s3a, ok := a.(*artifacts.S3Artifact); ok {
+			s3a.RawContentFile = ""
+		}
+	}
+
+	if !reflect.DeepEqual(got, expected) {
+		t.Fatalf("Expected different artifacts to be generated...\nExpected:\n%q\nActual:\n%q", expected, got)
 	}
 }
 
@@ -61,9 +70,9 @@ func TestFileArtifactWithNames(t *testing.T) {
 		},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/build/firefox.exe",
 					Expires: inAnHour,
 				},
@@ -91,9 +100,9 @@ func TestFileArtifactWithContentType(t *testing.T) {
 		},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/build/firefox.exe",
 					Expires: inAnHour,
 				},
@@ -159,9 +168,9 @@ func TestFileArtifactWithContentEncoding(t *testing.T) {
 		},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -169,8 +178,8 @@ func TestFileArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            "SampleArtifacts/_/X.txt",
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -178,8 +187,8 @@ func TestFileArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "identity",
 				Path:            "SampleArtifacts/b/c/d.jpg",
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -187,8 +196,8 @@ func TestFileArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "identity",
 				Path:            "SampleArtifacts/_/X.txt",
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -196,8 +205,8 @@ func TestFileArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "identity",
 				Path:            "SampleArtifacts/b/c/d.jpg",
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -205,8 +214,8 @@ func TestFileArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            "SampleArtifacts/_/X.txt",
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -233,9 +242,9 @@ func TestDirectoryArtifactWithNames(t *testing.T) {
 		},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/%%%/v/X",
 					Expires: inAnHour,
 				},
@@ -243,8 +252,8 @@ func TestDirectoryArtifactWithNames(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "%%%", "v", "X"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -252,8 +261,8 @@ func TestDirectoryArtifactWithNames(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "_", "X.txt"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -281,9 +290,9 @@ func TestDirectoryArtifactWithContentType(t *testing.T) {
 		},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/%%%/v/X",
 					Expires: inAnHour,
 				},
@@ -291,8 +300,8 @@ func TestDirectoryArtifactWithContentType(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "%%%", "v", "X"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -300,8 +309,8 @@ func TestDirectoryArtifactWithContentType(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "_", "X.txt"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -338,9 +347,9 @@ func TestDirectoryArtifactWithContentEncoding(t *testing.T) {
 		},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/%%%/v/X",
 					Expires: inAnHour,
 				},
@@ -348,8 +357,8 @@ func TestDirectoryArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "identity",
 				Path:            filepath.Join("SampleArtifacts", "%%%", "v", "X"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -357,8 +366,8 @@ func TestDirectoryArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "identity",
 				Path:            filepath.Join("SampleArtifacts", "_", "X.txt"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -366,8 +375,8 @@ func TestDirectoryArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "identity",
 				Path:            filepath.Join("SampleArtifacts", "b", "c", "d.jpg"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/%%%/v/X",
 					Expires: inAnHour,
 				},
@@ -375,8 +384,8 @@ func TestDirectoryArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "%%%", "v", "X"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -384,8 +393,8 @@ func TestDirectoryArtifactWithContentEncoding(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "_", "X.txt"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "public/b/c/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -413,9 +422,9 @@ func TestDirectoryArtifacts(t *testing.T) {
 		}},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "SampleArtifacts/%%%/v/X",
 					Expires: inAnHour,
 				},
@@ -423,8 +432,8 @@ func TestDirectoryArtifacts(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "%%%", "v", "X"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "SampleArtifacts/_/X.txt",
 					Expires: inAnHour,
 				},
@@ -432,8 +441,8 @@ func TestDirectoryArtifacts(t *testing.T) {
 				ContentEncoding: "gzip",
 				Path:            filepath.Join("SampleArtifacts", "_", "X.txt"),
 			},
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "SampleArtifacts/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -458,9 +467,9 @@ func TestMissingFileArtifact(t *testing.T) {
 		}},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&ErrorArtifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.ErrorArtifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    t.Name() + "/no_such_file",
 					Expires: inAnHour,
 				},
@@ -485,9 +494,9 @@ func TestMissingDirectoryArtifact(t *testing.T) {
 		}},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&ErrorArtifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.ErrorArtifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    t.Name() + "/no_such_dir",
 					Expires: inAnHour,
 				},
@@ -512,9 +521,9 @@ func TestFileArtifactIsDirectory(t *testing.T) {
 		}},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&ErrorArtifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.ErrorArtifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "SampleArtifacts/b/c",
 					Expires: inAnHour,
 				},
@@ -538,9 +547,9 @@ func TestDefaultArtifactExpiry(t *testing.T) {
 		}},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&S3Artifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.S3Artifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "SampleArtifacts/b/c/d.jpg",
 					Expires: inAnHour,
 				},
@@ -567,9 +576,9 @@ func TestDirectoryArtifactIsFile(t *testing.T) {
 		}},
 
 		// what we expect to discover on file system
-		[]TaskArtifact{
-			&ErrorArtifact{
-				BaseArtifact: &BaseArtifact{
+		[]artifacts.TaskArtifact{
+			&artifacts.ErrorArtifact{
+				BaseArtifact: &artifacts.BaseArtifact{
 					Name:    "SampleArtifacts/b/c/d.jpg",
 					Expires: inAnHour,
 				},
